Validación de Controladores, manejo de excepciones.

+++ Devsu/ApiDevsuCMunoz/ApiDevsuCMunoz.Application/Features/Cuentas/Commands/UpdateCuentas/UpdateCuentasCommand.cs
@@ -6,7 +6,7 @@ namespace ApiDevsuCMunoz.Application.Features.Cuentas.Commands.UpdateCuentas
     {
         public long Numero { get; set; }
         public string Tipo { get; set; } = string.Empty;
-        public double? SaldoInicial { get; set; }
+        public decimal? SaldoInicial { get; set; }
         public long? ClienteId { get; set; }
         public string Estado { get; set; } = string.Empty;
     }
