+++ Devsu/ApiDevsuCMunoz/ApiDevsuCMunoz.Application/Features/Cuentas/VModels/CuentasVM.cs
@@ -10,7 +10,7 @@ namespace ApiDevsuCMunoz.Application.Features.Cuentas.VModels
     {
         public long Numero { get; set; }
         public string? Tipo { get; set; } = string.Empty;
-        public double? SaldoInicial { get; set; }
+        public decimal? SaldoInicial { get; set; }
         public long? ClienteId { get; set; }
         public string? Estado { get; set; } = string.Empty;
     }
