+++ Devsu/ApiDevsuCMunoz/ApiDevsuCMunoz.Domain/Cuenta.cs
@@ -7,7 +7,7 @@ namespace ApiDevsuCMunoz.Domain
         [Key]
         public long Numero { get; set; }
         public string? Tipo { get; set; }
-        public double? SaldoInicial { get; set; }
+        public decimal? SaldoInicial { get; set; }
         public long? ClienteId { get; set; }
         public string? Estado { get; set; }
         public virtual Cliente? Cliente { get; set; }
